@@ -34,23 +34,11 @@
 		</div>
 	</div>
 
-		<div class="popup">
-			<div class="popup-card">
-				<button class="button-close" type="button">X</button>
-				<h2>Выберите ваш город</h2>
-				<ul>
-					<li><button class="query-button" type="button" data-name="Дзержинск" data-mail="autoakula@mail.ru" data-vk="https://vk.com/avtozapchastidriverdzerzhinsk">Дзержинск</button></li>
-					<li><button class="query-button" type="button" data-name="Слуцк" data-mail="driver.slutsk@yandex.by" data-vk="https://m.vk.com/club137761341">Слуцк</button></li>
-					<li><button class="query-button" type="button" data-name="Старые Дороги" data-mail="driver.stdorogi@yandex.by" data-vk="https://m.vk.com/club137761341">Старые Дороги</button></li>
-				</ul>
-			</div>
-		</div>
-
 		<header class="page-header" id="menu">
 			<div style="text-align: center;margin: auto 0 auto 8%;"> 
 				<img src="~/Images/driver-logo-e1527517042907.png" class="site-logo">
 				<div  class="info-city">
-					<button class="open-popup" type="button">Изменить город</button>
+					<button class="open-popup" type="button">Дзержинск</button>
 				</div>
 			</div>
 				<div class="phone-button menu-open click" id="menu-open"><span></span></div>
@@ -75,7 +63,7 @@
     		<div class="footer-container">
     			<div class="footer-logo">
     				<img src="~/Images/driver-logo-e1527517042907.png">
-    				<p><span>&#169</span>2019. Все права защищены.</P>
+    				<p><span>&#169</span>2024. Все права защищены.</P>
     			</div>
     			<div class="footer-info">
     				<div class="social">
@@ -100,7 +88,6 @@
     	});
     	</script>
     	<script src="~/js/preloader.js"></script>
-    	<script src="~/js/city.js"></script>
     	<script src="~/js/menu.js"></script>
     	<script src="~/js/slider.js"></script>
     	<script src="~/js/wow.min.js"></script>
