@@ -1,6 +1,9 @@
 //------------------------------------------------------------------------------
 // <auto-generated>
-//     Generated by the MSBuild WriteCodeFragment class.
+//     This code was generated by a tool.
+//
+//     Changes to this file may cause incorrect behavior and will be lost if
+//     the code is regenerated.
 // </auto-generated>
 //------------------------------------------------------------------------------
 
@@ -8,3 +11,6 @@ using System;
 using System.Reflection;
 
 [assembly: Microsoft.AspNetCore.Mvc.ApplicationParts.RelatedAssemblyAttribute("Driver.Views")]
+
+// Generated by the MSBuild WriteCodeFragment class.
+
